+++ src/layers/tinker.hpp
@@ -2,7 +2,6 @@
 
 namespace uiscaling::tinker {
     geode::Mod* get();
-    bool scalingEnabled();
     /// trick tinker into making this the ui scale
     void updateUI(float pScale, bool pScaleToolbar, bool pUseSafeArea, std::optional<float> pCustomSafeArea);
     /// fix tinker state